# packages/portal-contracts/openapi/portal-api-v1.json
{
  "components": {
    "schemas": {
      "CapabilityListView": {
        "additionalProperties": false,
        "properties": {
          "capabilities": {
            "items": {
              "$ref": "#/components/schemas/CapabilityView"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "capability_revision": {
            "title": "Capability Revision",
            "type": "string"
          }
        },
        "required": [
          "capabilities",
          "capability_revision"
        ],
        "title": "CapabilityListView",
        "type": "object"
      },
      "CapabilityView": {
        "additionalProperties": false,
        "properties": {
          "capability_id": {
            "title": "Capability Id",
            "type": "string"
          },
          "environment_id": {
            "title": "Environment Id",
            "type": "string"
          },
          "mode": {
            "title": "Mode",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "capability_id",
          "environment_id",
          "mode",
          "state"
        ],
        "title": "CapabilityView",
        "type": "object"
      },
...
      "EnvironmentListView": {
        "additionalProperties": false,
        "properties": {
          "capability_revision": {
            "title": "Capability Revision",
            "type": "string"
          },
          "environments": {
            "items": {
              "$ref": "#/components/schemas/EnvironmentView"
            },
            "title": "Environments",
            "type": "array"
          },
          "policy_revision": {
            "title": "Policy Revision",
            "type": "string"
          }
        },
        "required": [
          "environments",
          "policy_revision",
          "capability_revision"
        ],
        "title": "EnvironmentListView",
        "type": "object"
      },
      "EnvironmentSelectionRequest": {
        "additionalProperties": false,
        "properties": {
          "environment_id": {
            "title": "Environment Id",
            "type": "string"
          }
        },
        "required": [
          "environment_id"
        ],
        "title": "EnvironmentSelectionRequest",
        "type": "object"
      },
      "EnvironmentSelectionView": {
        "additionalProperties": false,
        "properties": {
          "capability_revision": {
            "title": "Capability Revision",
            "type": "string"
          },
          "environment_id": {
            "title": "Environment Id",
            "type": "string"
          },
          "policy_revision": {
            "title": "Policy Revision",
            "type": "string"
          }
        },
        "required": [
          "environment_id",
          "policy_revision",
          "capability_revision"
        ],
        "title": "EnvironmentSelectionView",
        "type": "object"
      },
      "EnvironmentView": {
        "additionalProperties": false,
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "environment_id": {
            "title": "Environment Id",
            "type": "string"
          },
          "selected": {
            "default": false,
            "title": "Selected",
            "type": "boolean"
          }
        },
        "required": [
          "environment_id",
          "display_name"
        ],
        "title": "EnvironmentView",
        "type": "object"
      },
      "FieldError": {
        "properties": {
          "field": {
            "title": "Field",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "message"
        ],
        "title": "FieldError",
        "type": "object"
      },
...
      "NavigationItem": {
        "additionalProperties": false,
        "properties": {
          "capability_id": {
            "title": "Capability Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "label",
          "path",
          "capability_id"
        ],
        "title": "NavigationItem",
        "type": "object"
      },
      "NavigationView": {
        "additionalProperties": false,
        "properties": {
          "capability_revision": {
            "title": "Capability Revision",
            "type": "string"
          },
          "environment_id": {
            "title": "Environment Id",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/NavigationItem"
            },
            "title": "Items",
            "type": "array"
          },
          "policy_revision": {
            "title": "Policy Revision",
            "type": "string"
          }
        },
        "required": [
          "environment_id",
          "items",
          "policy_revision",
          "capability_revision"
        ],
        "title": "NavigationView",
        "type": "object"
      },
      "ProblemDetails": {
        "additionalProperties": false,
        "description": "Portal-wide error response contract.",
        "properties": {
          "correlation_id": {
            "title": "Correlation Id",
            "type": "string"
          },
          "dependency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dependency"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "documentation_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Documentation Url"
          },
          "error_code": {
            "title": "Error Code",
            "type": "string"
          },
          "field_errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field Errors"
          },
          "instance": {
            "title": "Instance",
            "type": "string"
          },
          "retry_after_seconds": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry After Seconds"
          },
          "retryable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retryable"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "error_code",
          "correlation_id",
          "timestamp"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
...
    }
  },
  "info": {
    "contact": {
      "name": "Data Platform Engineering"
    },
    "description": "PR-PORTAL-001 exposes safe foundation health and system metadata only. No data-platform business capabilities or mutations are enabled.",
    "license": {
      "name": "Private repository"
    },
    "summary": "Foundation BFF for the Enterprise Data Platform Portal",
    "title": "Fintech Data Platform Portal API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/capabilities": {
      "get": {
        "operationId": "listCapabilities",
        "parameters": [
          {
            "in": "query",
            "name": "environment_id",
            "required": true,
            "schema": {
              "title": "Environment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilityListView"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Resource not found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Method not allowed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid request"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Sanitized internal error"
          }
        },
        "summary": "List Capabilities",
        "tags": [
          "access"
        ]
      }
    },
    "/v1/environments": {
      "get": {
        "operationId": "listEnvironments",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvironmentListView"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Resource not found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Method not allowed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid request"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Sanitized internal error"
          }
        },
        "summary": "List Environments",
        "tags": [
          "access"
        ]
      }
    },
    "/v1/navigation": {
      "get": {
        "operationId": "getNavigation",
        "parameters": [
          {
            "in": "query",
            "name": "environment_id",
            "required": true,
            "schema": {
              "title": "Environment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NavigationView"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Resource not found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Method not allowed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid request"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Sanitized internal error"
          }
        },
        "summary": "Get Navigation",
        "tags": [
          "access"
        ]
      }
    },
...
    "/v1/session/environment": {
      "post": {
        "operationId": "selectEnvironment",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnvironmentSelectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvironmentSelectionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Resource not found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Method not allowed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid request"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Sanitized internal error"
          }
        },
        "summary": "Select Environment",
        "tags": [
          "access"
        ]
      }
    },
...
  }
}